Major re-jigging
Moved the remaining business logic out of the controllers, and into
their own models, PatientHandler and MedicineHandler. Now it's very MVC,
which is good.

Added a little JSPF which displays ErrorBeans if they exist.

Added a new WebServices page and Servlet, for the eventual web service.

Some other minor changes, like the modification of the Patients,
Medicine and ErrorBean constructors.

Most of the test cases pass, except for the ones related to patients.
This appears to be due to some dodgy SQL on my part. I'll fix this
tomorrow. Once that's done, all we need to do is add the web-services,
and possibly think of another interesting way to use filters, and we'll
be done!