Forgive me Father, for I have sinned.
Made a massive, MASSIVE change to the system. Turns out the way I was
doing the MVC was wrong, according to my course leader. What we're
supposed to do is have one front controller which basically just does
routing and login stuff.

This means that all the models are no longer called from the
controller(s), but in the jsp view themselves. In a scriptlet. URGH.

On the plus side, everything still works; the site is still functionally
correct.

In other news, I made a filter that grabs information about the client
(and added the new database dump with the table that facilitates this),
implemented some session caching, and made a start on the web services
stuff.

All that's needed now are some interaction diagrams, the web service
stuff, and some more comments and we're gravy.