Finished!
Added web service stuff - a spell checker. Commented about as much as I
think is necessary - methods and stuff, and some lines that might need
explaining. Added 'interaction' diagrams that describe how the system
works for a couple of examples - patient viewing, logging in, etc.

All that need to be done now is test it at uni, and hand it in.