Few minor fixes
Changed some stuff. The select method in patient handler now returns a
list of maps, which give you the same basic functionality as the
ResultSet, except it's not tied to the database, and is thus more
stable.

Explicitly set the scope to request when storing lists/objects in the
scriplets. Otherwise, the PageContext is quite happy just deleting them,
because it wants to make my life slightly more difficult.

Stored a DatabaseHandler in the ServletContext instead of a Connection.
This saves literally one line of code in each class that uses the dbh.
Yay?

Finally, need to find a way to add varying quantities of medicines to a
patient, and test that this all works on Uni systems. It was being a bit
funny today